feat: extract create schema modal to separate component
